+++ TicketCenterL/src/main/java/bg/tuvarna/si/ticketcenterl/repositories/EventRepository.java
@@ -7,13 +7,13 @@ import org.springframework.stereotype.Repository;
 import java.util.Date;
 import java.util.List;
 @Repository
-public interface EventRepository extends JpaRepository<Event, Long> {
-    Event findEventById(Long eventId);
+public interface EventRepository extends JpaRepository<Event, Integer> {
+    Event findEventById(Integer eventId);
 
     List<Event> findByName(String name);
-    List<Event> findByDistributorsId(Long distributorId);
+    List<Event> findByDistributorsId(Integer distributorId);
     List<Event> findByStatus(String status);
     List<Event> findByDate(Date date);
 
-    List<Event> findByOrganizerId(Long organizerId);
+    List<Event> findByOrganizerId(Integer organizerId);
 }
